+++ changelog.md
@@ -6,6 +6,8 @@
 * fn:indexof() now works for both strings and arrays
 * fn:trimindent() to detect and remove constant number of spaces from lines
 * fn:replace(), fn:replaceall() for string replacement (no regex support yet)
+* add //@hibiki/confirm() and //@hibiki/alert()
+* rename //@hibiki/setInterval and //@hibiki/setTimeout to //@hibiki/set-interval and //@hibiki/set-timeout (old names still work).  add //@hibiki/clear-timeout
 * special syntax to turn html comments to hibiki text nodes using "hibiki:text" or "hibiki:rawtext"
 * bugfix: fix isnoattr() to not throw errors